refactor(devtools): relocate daemon_workload_probe + archive_space_report into polylogue
Both modules are production code that happened to live under devtools/:
polylogue/cli/commands/diagnostics.py's `polylogue ops diagnostics
workload`/`space` commands import them directly, and neither module
itself imports anything from devtools. Moves them to
polylogue/operations/ (matching that package's existing flat-file
layout and its role as high-level archive operations), moves their
test files to tests/unit/operations/ to match, and updates every
import site: the CLI command module, devtools/degraded_archive_proof.py
(a devtools-internal consumer, untouched otherwise), the integration
test, the benchmark test, and one docs/internals.md path reference.
Also fixes one now-stale 'devtools build' error string inside
daemon_workload_probe.py's version-mismatch message.

No CommandSpec entries existed for either module in
devtools/command_catalog.py -- they were never devtools commands, only
implementation the native polylogue CLI happened to import lazily from
the wrong package.

Ref .agent/campaigns/2026-08-overhaul/ws-c-devtools-pruning.md (Tranche 3)
